% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asinh_trans.R
\name{asinh_trans}
\alias{asinh_trans}
\title{Inverse Hyperbolic Sine Transformation for ggplot2}
\usage{
asinh_trans()
}
\value{
A transformation object for use with ggplot2 scale functions
}
\description{
Creates a transformation object for use with ggplot2 scales that applies
inverse hyperbolic sine (asinh) transformation, commonly used for flow cytometry data.
}
\examples{
\dontrun{
library(ggplot2)
ggplot(data, aes(x = FSC, y = SSC)) +
  geom_point() +
  scale_x_continuous(trans = asinh_trans()) +
  scale_y_continuous(trans = asinh_trans())
}
}
