% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fcs_exprs.R
\name{extract_fcs_exprs}
\alias{extract_fcs_exprs}
\title{Extract Expression Data from flowSet}
\usage{
extract_fcs_exprs(flowset)
}
\arguments{
\item{flowset}{A flowSet object containing flow cytometry data}
}
\value{
A tibble containing expression data from all samples with fileName and cell_ID columns
}
\description{
Converts a flowSet object to a tibble containing expression data with sample names and cell IDs.
}
\examples{
\dontrun{
# Assuming you have a flowSet object called fs
expr_data <- extract_fcs_exprs(fs)
}
}
